backend/imager/requirements.txt
python-socketio[client]==4.3.1
opencv-python==4.1.1.26
numpy==1.17.0